requirements.txt
flask
easyocr
opencv-python
pillow
pdf2image
pandas
numpy
scikit-learn